fix(tools): relocate skills toolset to a worker-importable lib
Agents execute in the Temporal worker, but the agentarea/skills toolset lived
in agentarea_api.tools (the API app), which the worker cannot import
(ModuleNotFoundError: No module named 'agentarea_api'). The code-tools loader
caught that and only debug-logged, so agentarea/skills was silently absent at
agent runtime: equipping an agent with it via the API was a no-op, and the
agent correctly reported 'I don't have a skills import tool'.

Move skills_toolset.py and its thin context base (base.py -> platform_base.py,
pure agentarea_common + SDK, ContextVar-based, no api-app deps) into
agentarea_agents.tools (a worker lib), point the loader at the new path, and
leave re-export shims at the old agentarea_api.tools paths so MCP
get_platform_tools, client_mcp and the sibling toolsets keep working. Repoint
the github-import test to the new module so its monkeypatches hit the real code.

Verified end-to-end on the live Kimi agent: after equipping agentarea/skills,
the agent calls import_from_github and the skill lands in the S3 skill store
(source_url + real id). The other ~13 agentarea_api.tools.* platform toolsets
remain API-only until similarly relocated.
