+++ agentarea-platform/libs/agents/agentarea_agents/tools/__init__.py
@@ -1,0 +1,5 @@
+"""Platform toolsets that agents use at runtime (worker-importable).
+
+Relocated from apps/api so the Temporal worker (which cannot import
+``agentarea_api``) can register them. apps/api keeps thin re-export shims.
+"""
